pyproject.toml
[tool.poetry]
name = "aaa"
version = "0.1.0"
description = ""
authors = ["Сергей Сушков <sushkov_sv@ursip.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
psutil = "^5.9.6"
pandas = "^2.1.2"
openpyxl = "^3.1.2"
xlsxwriter = "^3.1.9"
xlwings = "^0.30.12"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
